+++ cli/tests/pcluster/config/test_validators.py
@@ -83,6 +83,14 @@ def test_cluster_validator(mocker, section_dict, expected_message):
     "instance_type, expected_message", [("t2.micro", None), ("c4.xlarge", None), ("c5.xlarge", "is not supported")]
 )
 def test_ec2_instance_type_validator(mocker, instance_type, expected_message):
+    config_parser_dict = {"cluster default": {"compute_instance_type": instance_type}}
+    utils.assert_param_validator(mocker, config_parser_dict, expected_message)
+
+
+@pytest.mark.parametrize(
+    "instance_type, expected_message", [("t2.micro", None), ("c4.xlarge", None), ("p4d.24xlarge", "is not supported")]
+)
+def test_head_node_instance_type_validator(mocker, instance_type, expected_message):
     config_parser_dict = {"cluster default": {"master_instance_type": instance_type}}
     utils.assert_param_validator(mocker, config_parser_dict, expected_message)