+++ CHANGELOG.md
@@ -9,10 +9,10 @@ CHANGELOG
 - Add support for CentOS 8 in all Commercial regions.
 - Enable support for NICE DCV in GovCloud regions.
 - Enable support for AWS Batch scheduler in GovCloud regions.
-- Add support for P4d instance type 
+- Add support for P4d instance type as compute node.
 - Add ``-r/-region`` arg to ``pcluster configure``. If this arg is provided, configuration will skip region selection.
 - Add `cluster_resource_bucket` parameter under `cluster` section of cluster config to allow using user-provided bucket 
-  for hosting S3 resources used by cluster
+  for hosting S3 resources used by cluster.
 
 **CHANGES**